implement apartment and complex rest linking

+++ src/main/java/me/mirat1618/apartmentsapi/model/Apartment.java
@@ -48,7 +48,7 @@ public class Apartment {
     @JoinColumn(name = "apartment_complex_id")
     private ApartmentComplex apartmentComplex;
 
-    public Apartment(int number, int floor, int squareMeters, int numberOfBedrooms, boolean hasDishwasher, boolean hasWashingMachine, float rentPrice, ApartmentComplex apartmentComplex) {
+    public Apartment(int number, int floor, float squareMeters, int numberOfBedrooms, boolean hasDishwasher, boolean hasWashingMachine, float rentPrice, ApartmentComplex apartmentComplex) {
         this.number = number;
         this.squareMeters = squareMeters;
         this.numberOfBedrooms = numberOfBedrooms;
